Potentially fix bug in genDecl.go
I believe i64 is being moved into u64 even when u64 may already already be set (i.e. when isInt is false and isUint is true)

It seems like the intended behavior is to set u64 only when it's *not* set, which is when isUint is false and isInt true, i.e. when i64 is set instead of u64.